refactor: 다이어트 폼 리팩토링
