+++ src/app/(providers)/(root)/diets/write/_components/DietForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DietForm';
